requirements.txt
fastapi
uvicorn
pandas
numpy
openai
python-dotenv
pydantic
email-validator
requests
chromadb
python-jose[cryptography]
passlib[bcrypt]
sqlalchemy
psycopg2-binary
python-multipart
